+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2023-02-11
+### Added
+- Support for Calculation Groups.
+- Support for Privacy Levels in Data Sources.
+- Support for PostgresSQL Data Sources
+- Support for MySQL Data Sources.
+
+
 ## [0.1.2] - 2023-01-03
 ### Fixed
 - Columns not including `sourceColumn` and `sortByColumn` invalidating the bim file.